+++ lib/lib_basic/IRremoteESP8266/IRremoteESP8266/src/ir_Argo.cpp
@@ -344,7 +344,7 @@ stdAc::fanspeed_t IRArgoAC::toCommonFanSpeed(const uint8_t speed) {
 /// Convert the current internal state into its stdAc::state_t equivalent.
 /// @return The stdAc equivalent of the native settings.
 stdAc::state_t IRArgoAC::toCommon(void) const {
-  stdAc::state_t result;
+  stdAc::state_t result{};
   result.protocol = decode_type_t::ARGO;
   result.power = _.Power;
   result.mode = toCommonMode(_.Mode);