+++ lib/lib_basic/IRremoteESP8266/IRremoteESP8266/src/ir_Corona.cpp
@@ -550,7 +550,7 @@ String IRCoronaAc::toString(void) const {
 /// Convert the A/C state to it's common stdAc::state_t equivalent.
 /// @return A stdAc::state_t state.
 stdAc::state_t IRCoronaAc::toCommon() const {
-  stdAc::state_t result;
+  stdAc::state_t result{};
   result.protocol = decode_type_t::CORONA_AC;
   result.model = -1;  // No models used.
   result.power = _.Power;