+++ lib/lib_basic/IRremoteESP8266/IRremoteESP8266/src/ir_Delonghi.cpp
@@ -421,7 +421,7 @@ uint16_t IRDelonghiAc::getOffTimer(void) const {
 /// Convert the current internal state into its stdAc::state_t equivalent.
 /// @return The stdAc equivalent of the native settings.
 stdAc::state_t IRDelonghiAc::toCommon(void) const {
-  stdAc::state_t result;
+  stdAc::state_t result{};
   result.protocol = decode_type_t::DELONGHI_AC;
   result.power = _.Power;
   // result.mode = toCommonMode(getMode());